+++ crates/rust-client/src/rpc/tonic_client/mod.rs
@@ -368,6 +368,10 @@ impl NodeRpcClient for GrpcClient {
     /// updated to use the new commitment on subsequent requests. If the client is not connected,
     /// the commitment will be stored and used when the client connects. If the genesis commitment
     /// is already set, this method does nothing.
+    fn has_genesis_commitment(&self) -> Option<Word> {
+        *self.genesis_commitment.read()
+    }
+
     async fn set_genesis_commitment(&self, commitment: Word) -> Result<(), RpcError> {
         // Check if already set before doing anything else
         if self.genesis_commitment.read().is_some() {
@@ -965,6 +969,10 @@ impl NodeRpcClient for GrpcClient {
         // Cache fetched values
         self.limits.write().replace(limits);
         Ok(limits)
+    }
+
+    fn has_rpc_limits(&self) -> Option<RpcLimits> {
+        *self.limits.read()
     }
 
     async fn set_rpc_limits(&self, limits: RpcLimits) {